pyproject.toml
[project]
name = "facedetectionpy"
version = "0.1.0"
description = ""
authors = [
    {name = "yukesh570",email = "yukesh.maha10@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "absl-py (==2.1.0)",
    "asgiref (==3.8.1)",
    "astunparse (==1.6.3)",
    "beautifulsoup4 (==4.12.3)",
    "blinker (==1.8.2)",
    "certifi (==2024.7.4)",
    "charset-normalizer (==3.3.2)",
    "click (==8.1.7)",
    "colorama (==0.4.6)",
    "django (==5.1)",
    "djangorestframework (==3.15.2)",
    "filelock (==3.15.4)",
    "fire (==0.6.0)",
    "flask (==3.0.3)",
    "flask-cors (==4.0.1)",
    "flatbuffers (==24.3.25)",
    "gast (==0.6.0)",
    "gdown (==5.2.0)",
    "google-pasta (==0.2.0)",
    "gunicorn (==23.0.0)",
    "idna (==3.7)",
    "itsdangerous (==2.2.0)",
    "jinja2 (==3.1.4)",
    "keras (==3.5.0)",
    "libclang (==18.1.1)",
    "markdown (==3.7)",
    "markdown-it-py (==3.0.0)",
    "markupsafe (==2.1.5)",
    "mdurl (==0.1.2)",
    "mtcnn (==0.1.1)",
    "namex (==0.0.8)",
    "opencv-python (==4.10.0.84)",
    "opt-einsum (==3.3.0)",
    "packaging (==24.1)",
    "pillow (==10.4.0)",
    "protobuf (==4.25.4)",
    "pygments (==2.18.0)",
    "pysocks (==1.7.1)",
    "python-dateutil (==2.9.0.post0)",
    "pytz (==2024.1)",
    "requests (==2.32.3)",
    "rich (==13.7.1)",
    "setuptools (==73.0.0)",
    "six (==1.16.0)",
    "soupsieve (==2.6)",
    "sqlparse (==0.5.1)",
    "tqdm (==4.66.5)",
    "typing-extensions (==4.12.2)",
    "tzdata (==2024.1)",
    "urllib3 (==2.2.2)",
    "werkzeug (==3.0.3)",
    "wheel (==0.44.0)",
    "wrapt (==1.16.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
